% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorities.R
\name{authorities}
\alias{authorities}
\alias{authority_names}
\alias{authority_years}
\title{authority parsing}
\usage{
authority_names(x)

authority_years(x)
}
\arguments{
\item{x}{character string of form: \code{<author>, <year>}}
}
\description{
authority parsing
}
\examples{
authority_names("Linnaeus, 1758")
authority_years("Linnaeus, 1758")

x <- c('Arrhenatherum elatius var. elatius',
           'Secale cereale subsp. cereale', 'Secale cereale ssp. cereale',
           'Vanessa atalanta (Linnaeus, 1758)')
           
x <- c('Smith, 1456', 'Foobarcheesestuff, 1001', 'Smith, 1987')
authority_names(x)
authority_years(x)
}
