% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_parse.R
\name{pgx_parse}
\alias{pgx_parse}
\title{parse complete names}
\usage{
pgx_parse(x)
}
\arguments{
\item{x}{scientific name character string}
}
\description{
parse complete names
}
\examples{
pgx_parse("Linnaeus")
pgx_parse("(Linnaeus)")
pgx_parse("Linnaeus 1789")
pgx_parse("Helianthus annuus L.")
\dontrun{
# needs a name
pgx_parse("1789")
}
}
