% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_rank.R
\name{pgx_ranks}
\alias{pgx_ranks}
\title{parse out rank names}
\usage{
pgx_ranks(x)
}
\arguments{
\item{x}{character string}
}
\description{
parse out rank names
}
\examples{
pgx_ranks("var")
pgx_ranks("var.")
pgx_ranks("subsp.")
pgx_ranks("ssp.")
pgx_ranks("ssp. ")
pgx_ranks(" var. ")
pgx_ranks(" var.")
pgx_ranks("var. ")
pgx_ranks("[var.]")

pgx_ranks(c("[var.]", "var.", "ssp"))
}
