% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_sciname.R
\name{pgx_sciname}
\alias{pgx_sciname}
\title{scientific name parsing}
\usage{
pgx_sciname(x)
}
\arguments{
\item{x}{character string of any of the forms:
\code{<genus>}, \code{<genus> <epithet>}, \code{<genus> <epithet> <rank> <subtaxon>}, etc.}
}
\description{
these functions only parse names alone - not intended to parse
full names with authorities, etc.
}
\examples{
# single name, higher level taxon
pgx_sciname("Poaceae")

# single name, genus
pgx_sciname("Helianthus")
pgx_sciname("Helianthus ")

# two names, genus and epithet
pgx_sciname("Helianthus annuus")
pgx_sciname("Helianthus  annuus")
pgx_sciname("Helianthus  annuus ")

# genus, epithet, rank, and infraspecific epithet
pgx_sciname("Helianthus annuus var texanus")
pgx_sciname("Helianthus annuus var. texanus")

# unknown epithet variations
pgx_sciname("Helianthus sp.")
pgx_sciname("Helianthus sp")
pgx_sciname("Alyxia reinwardti ssp")
pgx_sciname("Helianthus spp.")
pgx_sciname("Helianthus spp")
pgx_sciname("Acarinina aff. pentacamerata")
}
