% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_subsp.R
\name{pgx_subsp}
\alias{pgx_subsp}
\title{parse out subspecies ranks}
\usage{
pgx_subsp(x)
}
\arguments{
\item{x}{character string}
}
\description{
parse out subspecies ranks
}
\examples{
pgx_subsp("var")
pgx_subsp("var.")
pgx_subsp("subsp.")
pgx_subsp("ssp.")
pgx_subsp("ssp. ")
pgx_subsp(" var. ")
pgx_subsp(" var.")
pgx_subsp("var. ")
pgx_subsp("[var.]")

pgx_subsp(c("[var.]", "var.", "ssp"))
}
