% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx_scrub.R
\name{pgx_symbols}
\alias{pgx_symbols}
\title{Clean taxonomic names of symbols}
\usage{
pgx_symbols(x)
}
\arguments{
\item{x}{(character) one or more taxonomic names}
}
\value{
names in a vector
}
\description{
Clean taxonomic names of symbols
}
\details{
doesn't use PEGs, just regular C++
}
\examples{
pgx_symbols(x = "Ste>phanolaimus spartinae")
pgx_symbols("Ste*phanolaimus")
pgx_symbols("{Stephanolaimus spartinae}")
pgx_symbols("Stephanolaimus spartinae|\%")
}
